feature/mop_mapping: Added tests for MoP

+++ source/TransformationTesting/src/main/scala/com/kainos/enstar/TransformationUnitTesting/MethodOfPlacementUtils.scala
@@ -3,12 +3,12 @@ package com.kainos.enstar.TransformationUnitTesting
 import org.apache.spark.sql.Row
 
 /**
-  * Created by caoimheb on 08/12/2016.
-  */
+ * Created by caoimheb on 08/12/2016.
+ */
 object MethodOfPlacementUtils {
 
   def lookupBusinessTypeMapping( cols : Array[String] ) : Row = {
-    Row( cols( 0 ), cols( 1 ) )
+    Row( cols( 0 ).toInt, cols( 1 ) )
   }
 
   def methodOfPlacementMapping( cols : Array[String] ) : Row = {
